Corrected some comments
