+++ piyavkin.anton/T2/datastruct.hpp
@@ -12,14 +12,15 @@ namespace piyavkin
     unsigned long long key1;
     std::pair< long long, unsigned long long > key2;
     std::string key3;
-    bool operator<(DataStruct& data);
-    bool operator>(DataStruct& data);
-    bool operator<=(DataStruct& data);
-    bool operator>=(DataStruct& data);
-    bool operator==(DataStruct& data);
-    bool operator!=(DataStruct& data);
+    bool operator<(const DataStruct& data) const;
+    bool operator>(const DataStruct& data) const;
+    bool operator<=(const DataStruct& data) const;
+    bool operator>=(const DataStruct& data) const;
+    bool operator==(const DataStruct& data) const;
+    bool operator!=(const DataStruct& data) const;
   };
   std::istream& operator>>(std::istream& in, DataStruct& data);
   std::ostream& operator<<(std::ostream& out, const DataStruct& data);
+  void inputKey(std::istream& in, size_t numberKey, DataStruct& data);
 }
 #endif
