+++ piyavkin.anton/T2/delimeter.cpp
@@ -21,9 +21,10 @@ std::istream& piyavkin::operator>>(std::istream& in, DelimeterChar&& exp)
 
 std::istream& piyavkin::operator>>(std::istream& in, DelimeterString&& exp)
 {
-  for (size_t i = 0; i < exp.expected.size() && in; ++i)
+  size_t i = 0;
+  while (exp.expected[i] != '\0')
   {
-    in >> DelimeterChar{exp.expected[i]};
+    in >> DelimeterChar{exp.expected[i++]};
   }
   return in;
 }
