--- piyavkin.anton/T2/inputkey.hpp
@@ -1,9 +1,0 @@
-#ifndef INPUTKEY_HPP
-#define INPUTKEY_HPP
-#include "datastruct.hpp"
-
-namespace piyavkin
-{
-  void inputKey(std::istream& in, size_t numberKey, DataStruct& data);
-}
-#endif
